+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Unquoted attribute values are now correctly parsed.
+
 ## [0.3.1] - 2022-02-13
 
 ### Fixed